添加测试代码

+++ service/buffer.go
@@ -274,7 +274,7 @@ func (this *buffer) ReadFrom(r io.Reader) (int64, error) {
 		// Total message length is remlen + 1 (msg type) + m (remlen bytes)
 
 		remlen_ := int64(remlen)
-		total = remlen_ + 1 + m
+		total = remlen_ + int64(1) + int64(m)
 		Log.Infoc(func() string {
 			return fmt.Sprintf("remlen===n===totle: %d===%d===%d", remlen,n,total)
 		})
